+++ CHANGELOG.md
@@ -32,6 +32,7 @@
   - Do not modify referenced objects when creating `StyleLayers` from `CustomLayer` or `VectorLayer.options.style` see also (cddf2fc6a83444851914ed6619bc8d5ada8ce485)
   - Fix add layer with the correct beforeId on update `PaintProperty` workaround for [maplibre issue 3001](https://github.com/maplibre/maplibre-gl-js/issues/3001).
   - Fix workaround that tried to get beforeId after layer was removed. Now use workaround only if `map.getTerrain()` is not not null.
+  - Fix set `layout.visibility` in style of `VectorLayer` based on `VectorLayer.visible`.
 
 * **@dlr-eoc/map-cesium:**
   - Fix get entity.name for infoBox in none GeoJson layers